pyproject.toml
[project]
name = "A-Star-GPS-Road-Navigation-System"
version = "0.1.0"
description = "This GPS Road System will find the (hopefully) shortest path to a destination using the A* graph traversal algorithm.  The graph is designed to use data from [Open Street Map](https://www.openstreetmap.org)."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "geopandas>=1.1.1",
    "osmnx>=2.0.7",
    "pillow>=12.0.0",
    "pyrosm>=0.6.2",
    "shapely>=2.1.2",
]
